+++ src/config.rs
@@ -70,8 +70,8 @@ impl Default for Config {
                 max_results: 0,
                 show_hidden_files: true,
                 exclude: vec![
-                    "**/node_modules/**".into(),
-                    "**/.git/**".into(),
+                    "**/node_modules".into(),
+                    "**/.git".into(),
                     "**/.DS_Store".into(),
                 ],
                 base_path: vec!["${workspaceFolder}".into(), "${document}".into()],
@@ -126,7 +126,7 @@ mod tests {
         assert!(cfg.completion.show_hidden_files);
         assert_eq!(
             cfg.completion.exclude,
-            vec!["**/node_modules/**", "**/.git/**", "**/.DS_Store"]
+            vec!["**/node_modules", "**/.git", "**/.DS_Store"]
         );
         assert_eq!(
             cfg.completion.base_path,